+++ src/main/java/io/github/zeroone3010/yahueapi/domain/ApiInitializationError.java
@@ -9,24 +9,12 @@ public class ApiInitializationError {
     return type;
   }
 
-  public void setType(int type) {
-    this.type = type;
-  }
-
   public String getAddress() {
     return address;
   }
 
-  public void setAddress(String address) {
-    this.address = address;
-  }
-
   public String getDescription() {
     return description;
-  }
-
-  public void setDescription(String description) {
-    this.description = description;
   }
 
   @Override