+++ src/main/java/io/github/zeroone3010/yahueapi/domain/LightConfig.java
@@ -10,32 +10,16 @@ public class LightConfig {
     return archetype;
   }
 
-  public void setArchetype(String archetype) {
-    this.archetype = archetype;
-  }
-
   public String getFunction() {
     return function;
-  }
-
-  public void setFunction(String function) {
-    this.function = function;
   }
 
   public String getDirection() {
     return direction;
   }
 
-  public void setDirection(String direction) {
-    this.direction = direction;
-  }
-
   public Startup getStartup() {
     return startup;
-  }
-
-  public void setStartup(Startup startup) {
-    this.startup = startup;
   }
 
   @Override