+++ src/main/java/io/github/zeroone3010/yahueapi/domain/LightState.java
@@ -36,132 +36,44 @@ public class LightState {
     return on;
   }
 
-  /**
-   * @param on
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setOn(boolean on) {
-    this.on = on;
-  }
-
   public int getBrightness() {
     return brightness;
-  }
-
-  /**
-   * @param brightness
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setBrightness(int brightness) {
-    this.brightness = brightness;
   }
 
   public int getHue() {
     return hue;
   }
 
-  /**
-   * @param hue
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setHue(int hue) {
-    this.hue = hue;
-  }
-
   public int getSaturation() {
     return saturation;
-  }
-
-  /**
-   * @param saturation
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setSaturation(int saturation) {
-    this.saturation = saturation;
   }
 
   public String getEffect() {
     return effect;
   }
 
-  /**
-   * @param effect
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setEffect(String effect) {
-    this.effect = effect;
-  }
-
   public List<Float> getXy() {
     return xy;
-  }
-
-  /**
-   * @param xy
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setXy(List<Float> xy) {
-    this.xy = xy;
   }
 
   public int getCt() {
     return ct;
   }
 
-  /**
-   * @param ct
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setCt(int ct) {
-    this.ct = ct;
-  }
-
   public String getAlert() {
     return alert;
-  }
-
-  /**
-   * @param alert
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setAlert(String alert) {
-    this.alert = alert;
   }
 
   public String getColorMode() {
     return colorMode;
   }
 
-  /**
-   * @param colorMode
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setColorMode(String colorMode) {
-    this.colorMode = colorMode;
-  }
-
   public String getMode() {
     return mode;
   }
 
-  /**
-   * @param mode
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setMode(String mode) {
-    this.mode = mode;
-  }
-
   public boolean isReachable() {
     return reachable;
-  }
-
-  /**
-   * @param reachable
-   * @deprecated The state cannot be changed with this class. Use {@link io.github.zeroone3010.yahueapi.State} instead.
-   */
-  public void setReachable(boolean reachable) {
-    this.reachable = reachable;
   }
 
   @Override