+++ src/main/java/io/github/zeroone3010/yahueapi/domain/PortalState.java
@@ -16,32 +16,16 @@ public class PortalState {
     return signedOn;
   }
 
-  public void setSignedOn(boolean signedOn) {
-    this.signedOn = signedOn;
-  }
-
   public boolean isIncoming() {
     return incoming;
-  }
-
-  public void setIncoming(boolean incoming) {
-    this.incoming = incoming;
   }
 
   public boolean isOutgoing() {
     return outgoing;
   }
 
-  public void setOutgoing(boolean outgoing) {
-    this.outgoing = outgoing;
-  }
-
   public String getCommunication() {
     return communication;
-  }
-
-  public void setCommunication(String communication) {
-    this.communication = communication;
   }
 
   @Override