+++ src/main/java/io/github/zeroone3010/yahueapi/domain/Root.java
@@ -28,64 +28,32 @@ public class Root {
     return lights;
   }
 
-  public void setLights(Map<String, LightDto> lights) {
-    this.lights = lights;
-  }
-
   public Map<String, Group> getGroups() {
     return groups;
-  }
-
-  public void setGroups(Map<String, Group> groups) {
-    this.groups = groups;
   }
 
   public BridgeConfig getConfig() {
     return config;
   }
 
-  public void setConfig(BridgeConfig config) {
-    this.config = config;
-  }
-
   public Map<String, Schedule> getSchedules() {
     return schedules;
-  }
-
-  public void setSchedules(Map<String, Schedule> schedules) {
-    this.schedules = schedules;
   }
 
   public Map<String, Scene> getScenes() {
     return scenes;
   }
 
-  public void setScenes(Map<String, Scene> scenes) {
-    this.scenes = scenes;
-  }
-
   public Map<String, Rule> getRules() {
     return rules;
-  }
-
-  public void setRules(Map<String, Rule> rules) {
-    this.rules = rules;
   }
 
   public Map<String, SensorDto> getSensors() {
     return sensors;
   }
 
-  public void setSensors(Map<String, SensorDto> sensors) {
-    this.sensors = sensors;
-  }
-
   public Map<String, ResourceLink> getResourcelinks() {
     return resourcelinks;
-  }
-
-  public void setResourcelinks(Map<String, ResourceLink> resourcelinks) {
-    this.resourcelinks = resourcelinks;
   }
 
   @Override