+++ src/main/java/io/github/zeroone3010/yahueapi/domain/SensorDto.java
@@ -34,96 +34,48 @@ public class SensorDto {
     return state;
   }
 
-  public void setState(Map<String, Object> state) {
-    this.state = state;
-  }
-
   public ComponentSoftwareUpdate getSoftwareUpdate() {
     return softwareUpdate;
-  }
-
-  public void setSoftwareUpdate(ComponentSoftwareUpdate softwareUpdate) {
-    this.softwareUpdate = softwareUpdate;
   }
 
   public Map<String, Object> getConfig() {
     return config;
   }
 
-  public void setConfig(Map<String, Object> config) {
-    this.config = config;
-  }
-
   public String getName() {
     return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
   }
 
   public String getType() {
     return type;
   }
 
-  public void setType(String type) {
-    this.type = type;
-  }
-
   public String getModelId() {
     return modelId;
-  }
-
-  public void setModelId(String modelId) {
-    this.modelId = modelId;
   }
 
   public String getManufacturerName() {
     return manufacturerName;
   }
 
-  public void setManufacturerName(String manufacturerName) {
-    this.manufacturerName = manufacturerName;
-  }
-
   public String getSoftwareVersion() {
     return softwareVersion;
-  }
-
-  public void setSoftwareVersion(String softwareVersion) {
-    this.softwareVersion = softwareVersion;
   }
 
   public String getUniqueId() {
     return uniqueId;
   }
 
-  public void setUniqueId(String uniqueId) {
-    this.uniqueId = uniqueId;
-  }
-
   public Boolean getRecycle() {
     return recycle;
-  }
-
-  public void setRecycle(Boolean recycle) {
-    this.recycle = recycle;
   }
 
   public String getProductName() {
     return productName;
   }
 
-  public void setProductName(String productName) {
-    this.productName = productName;
-  }
-
   public SensorCapabilities getCapabilities() {
     return capabilities;
-  }
-
-  public void setCapabilities(SensorCapabilities capabilities) {
-    this.capabilities = capabilities;
   }
 
   @Override